Changes after initial test run
using from . as a relative folder to import .py files didn't work so I qualified them

added a new parameter to run command --cumul-perc which we use to limit to top costing jobs. I will update readme for that

changed the way limit-row parameter works as None points to Limit None which fails the query. Added a very high number for now to pass the test runs. We can fix it

current flow creates "antipatterns" object as a part of the dictionary output of "job" class which prevents from pushing to BQ. I added a del function to remove it