A full refactor (#12)
* Fully Refactor

* Updated README.md

* Fixing stuff

* Antipatterns is now a class

* Fully fixed now

* Updated README.md

* Updated tests

* Changes after initial test run

using from . as a relative folder to import .py files didn't work so I qualified them

added a new parameter to run command --cumul-perc which we use to limit to top costing jobs. I will update readme for that

changed the way limit-row parameter works as None points to Limit None which fails the query. Added a very high number for now to pass the test runs. We can fix it

current flow creates "antipatterns" object as a part of the dictionary output of "job" class which prevents from pushing to BQ. I added a del function to remove it

* Update README.md

* Update antipattern-config.yaml

* Update README.md

* Fix problems with importing

* updated readme

* Fix the --limit-row

* Now accept multiple query_project and multiple information schema projects

* Update README.md

* adding new queries to handle multiple projects

* adding new queries to handle multiple projects

* Update functions.py

fixed cumul-perc variable

* Fixing Multiple day run

changing variables to run properly when days back parameter is more than 1 day

* Update jobs_query.sql.j2

* made union all prettier

* Update main.py

* Update functions.py

test to push multiple partitions

* Update functions.py

* Update antipattern-config.yaml and readme

* Update README.md

* Update config.py and main.py for multiple date run

* Update main.py

* Update main.py

---------

Co-authored-by: antoniodavide-cali <antoniodavide.cali@justeattakeaway.com>
Co-authored-by: Ozgur Uyar <ozgur.uyar@justeattakeaway.com>
Co-authored-by: Ozgur Sinan Uyar <ozguruyar@gmail.com>